move out FFTResolutionCombo
