+++ src/Features/PostProcessing/PostProcessingUI.h
@@ -1,0 +1,6 @@
+#pragma once
+
+namespace PostProcessingUI
+{
+	bool FFTResolutionCombo(const char* label, int& resolution);
+}
