+++ src/Utils/UI.cpp
@@ -59,10 +59,6 @@ namespace Util
 		IM_COL32(20, 240, 20, 255),
 		IM_COL32(20, 20, 240, 255)
 	};
-	constexpr std::array<const char*, 4> FFTResolutionLabels = { "128", "256", "512", "1024" };
-	constexpr std::array<int, FFTResolutionLabels.size()> FFTResolutionValues = { 128, 256, 512, 1024 };
-	constexpr int FFTResolutionDefaultIndex = 1;
-
 	void RefreshScreenScale(HWND hwnd, float bufferWidth, float bufferHeight)
 	{
 		RECT rect{};
@@ -440,28 +436,6 @@ namespace Util
 		}
 		ImGui::EndGroup();
 		return changed;
-	}
-
-	bool FFTResolutionCombo(const char* label, int& resolution)
-	{
-		int currentIndex = FFTResolutionDefaultIndex;
-		bool matched = false;
-		for (int i = 0; i < static_cast<int>(FFTResolutionValues.size()); i++) {
-			if (FFTResolutionValues[i] == resolution) {
-				currentIndex = i;
-				matched = true;
-				break;
-			}
-		}
-
-		if (!matched)
-			resolution = FFTResolutionValues[currentIndex];
-
-		if (!ImGui::Combo(label, &currentIndex, FFTResolutionLabels.data(), static_cast<int>(FFTResolutionLabels.size())))
-			return !matched;
-
-		resolution = FFTResolutionValues[currentIndex];
-		return true;
 	}
 
 	ImVec2 GetNativeViewportSizeScaled(float scale)
